Add files via upload
the delete bot is the code that deletes the images while the keras files are the AI I created that tells the difference between the images you give it. My code also deals with errors in batch downloading such as not being a png or jpg file and a file's character length being greater than an OS's max.